+++ app/Http/Controllers/Api/UserPostParticipationController.php
@@ -14,7 +14,6 @@ use App\Http\Controllers\Controller;
 use App\Models\User;
 use App\Models\Post;
 use App\Models\UserPointCategory;
-use Illuminate\Support\Facades\Cache;
 
 
 class UserPostParticipationController extends Controller
@@ -113,6 +112,7 @@ class UserPostParticipationController extends Controller
     public function update(Request $request, int $postId)
     {
         $user = $this->userService->getUser();
+        $this->authorize('update', $user);
         $validated = $request->validate([
             'is_completed' => 'boolean',
         ]);
@@ -129,6 +129,8 @@ class UserPostParticipationController extends Controller
      */
     public function destroy(int $postId, int $userId)
     {
+        $this->authorize('delete', $userId);
+
         $userPostParticipation = UserPostParticipation::where(['post_id' => $postId, 'participant_id' => $userId])->firstOrFail();
 
         $userPostParticipation->delete();
@@ -169,9 +171,7 @@ class UserPostParticipationController extends Controller
         if (!$userAuthenticated || !$user) {
             return response()->json(['error' => 'User not found.'], 404);
         }
-        if (!$this->userService->checkIfCanAccessToResource($userId) || !$this->userService->isUserUnblocked($user->id)) {
-            return response()->json(['error' => 'Access denied'], 403);
-        }
+        $this->authorize('view', $userId);
 
         $userPostParticipations = UserPostParticipation::where('participant_id', $user->id)->get();
         $userPostParticipations->load('posts');
@@ -183,15 +183,8 @@ class UserPostParticipationController extends Controller
      */
     public function getPostsByUserCompleted(int $userId)
     {
-        if (Cache::has('challenges_completed_user_' . $userId)) {
-            if ($this->userService->checkIfCanAccessToResource($userId) && $this->userService->isUserUnblocked($userId)) {
-                return response()->json(Cache::get('challenges_completed_user_' . $userId));
-            }
-        }
         $user = User::where('id', $userId)->firstOrFail();
-        if (!$this->userService->checkIfCanAccessToResource($userId) || !$this->userService->isUserUnblocked($user->id)) {
-            return response()->json(['error' => 'Access denied'], 403);
-        }
+        $this->authorize('view', $user);
 
         $userPostParticipations = UserPostParticipation::where(['participant_id' => $user->id, 'is_completed' => true])->get();
         $userPostParticipations->load('posts')->where('type', "challenge");
@@ -208,11 +201,6 @@ class UserPostParticipationController extends Controller
             }
         }
 
-        if ($this->userService->checkIfCanAccessToResource($userId) && $this->userService->isUserUnblocked($userId)) {
-            Cache::put('challenges_completed_user_' . $userId, $userChallenges, 60);
-        }
-
-
         return response()->json($userChallenges);
     }
 
@@ -221,12 +209,9 @@ class UserPostParticipationController extends Controller
      */
     public function getPostsByUserAbandoned(int $userId)
     {
-        
-        if (!$this->userService->checkIfCanAccessToResource($userId) || !$this->userService->isUserUnblocked($userId)) {
-            return response()->json(['error' => 'Access denied'], 403);
-        }
-
         $user = User::where('id', $userId)->firstOrFail();
+        $this->authorize('view', $user);
+
         $userPostParticipations = UserPostParticipation::where(['participant_id' => $user->id, 'is_completed' => false])->get();
         $userPostParticipations->load('posts')->where('type', "challenge");
 
@@ -252,16 +237,8 @@ class UserPostParticipationController extends Controller
      */
     public function getPostsByUserInProgress(int $userId)
     {
-
-        if (Cache::has('challenges_inprogress_user_' . $userId)) {
-            if ($this->userService->checkIfCanAccessToResource($userId) && $this->userService->isUserUnblocked($userId)) {
-                return response()->json(Cache::get('challenges_inprogress_user_' . $userId));
-            }
-        }
         $user = User::where('id', $userId)->firstOrFail();
-        if (!$this->userService->checkIfCanAccessToResource($userId) || !$this->userService->isUserUnblocked($user->id)) {
-            return response()->json(['error' => 'Access denied'], 403);
-        }
+        $this->authorize('view', $user);
 
         $userPostParticipations = UserPostParticipation::where(['participant_id' => $user->id, 'is_completed' => false])->get();
         $userPostParticipations->load('posts')->where('type', "challenge");
@@ -279,10 +256,6 @@ class UserPostParticipationController extends Controller
             }
         }
 
-        if ($this->userService->checkIfCanAccessToResource($userId) && $this->userService->isUserUnblocked($user->id)) {
-            Cache::put('challenges_inprogress_user_' . $userId, $userPostParticipationsInProgress, 60);
-        }
-
         return response()->json($userPostParticipationsInProgress);
     }
 
@@ -292,16 +265,8 @@ class UserPostParticipationController extends Controller
      */
     public function getPostsByUserNext(int $userId)
     {
-        if (Cache::has('challenges_next_user_' . $userId)) {
-            if ($this->userService->checkIfCanAccessToResource($userId) && $this->userService->isUserUnblocked($userId)) {
-                return response()->json(Cache::get('challenges_next_user_' . $userId));
-            }
-        }
-
         $user = User::where('id', $userId)->firstOrFail();
-        if (!$this->userService->checkIfCanAccessToResource($userId) || !$this->userService->isUserUnblocked($user->id)) {
-            return response()->json(['error' => 'Access denied'], 403);
-        }
+        $this->authorize('view', $user);
 
         $userPostParticipations = UserPostParticipation::where(['participant_id' => $user->id, 'is_completed' => false])->get();
         $userPostParticipations->load('posts')->where('type', "challenge");
@@ -321,11 +286,7 @@ class UserPostParticipationController extends Controller
                 }
             }
         }
-
-        if ($this->userService->checkIfCanAccessToResource($userId) && $this->userService->isUserUnblocked($user->id)) {
-            Cache::put('challenges_next_user_' . $userId, $userPostParticipationsNext, 60);
-        }
-
+        
         return response()->json($userPostParticipationsNext);
     }
 
@@ -334,15 +295,8 @@ class UserPostParticipationController extends Controller
      */
     public function getUserActions(int $userId)
     {
-        if (Cache::has('actions_user_' . $userId)) {
-            if ($this->userService->checkIfCanAccessToResource($userId) && $this->userService->isUserUnblocked($userId)) {
-                return response()->json(Cache::get('actions_user_' . $userId));
-            }
-        }
         $user = User::where('id', $userId)->firstOrFail();
-        if (!$this->userService->checkIfCanAccessToResource($userId) || !$this->userService->isUserUnblocked($user->id)) {
-            return response()->json(['error' => 'Access denied'], 403);
-        }
+        $this->authorize('view', $user);
 
         $userPostParticipations = UserPostParticipation::where(['participant_id' => $user->id, 'is_completed' => true])->get();
         $userActions = [];
@@ -357,10 +311,6 @@ class UserPostParticipationController extends Controller
 
                 $userActions[] = $post;
             }
-        }
-
-        if ($this->userService->checkIfCanAccessToResource($userId) && $this->userService->isUserUnblocked($user->id)) {
-            Cache::put('actions_user_' . $userId, $userActions, 60);
         }
 
         return response()->json($userActions);
